+++ lib/hexapdf/font/true_type_wrapper.rb
@@ -111,6 +111,11 @@ module HexaPDF
           false
         end
 
+        # Returns +true+ since this is a valid glyph.
+        def valid?
+          true
+        end
+
         #:nodoc:
         def inspect
           "#<#{self.class.name} font=#{@font_wrapper.wrapped_font.full_name.inspect} id=#{id} #{str.inspect}>"